sql update strategy
