+++ corehq/form_processor/backends/sql/update_strategy.py
@@ -1,0 +1,5 @@
+from corehq.form_processor.update_strategy_base import UpdateStrategy
+
+
+class SqlCaseUpdateStrategy(UpdateStrategy):
+    pass
